requirements.txt
pinecone
sentence-transformers
python-dotenv
numpy
